+++ tests/TestCase/DeprecationFunctionsTest.php
@@ -23,8 +23,8 @@ use Tools\TestSuite\TestCase;
 class DeprecationFunctionsTest extends TestCase
 {
     /**
-     * Test for `deprecationWarning()` global function
      * @test
+     * @uses \deprecationWarning()
      */
     public function testDeprecationWarning(): void
     {
@@ -32,9 +32,8 @@ class DeprecationFunctionsTest extends TestCase
         deprecationWarning('This method is deprecated');
         error_reporting($current);
 
-        $this->expectDeprecation();
-        $this->expectExceptionMessageMatches('/^This method is deprecated/');
-        $this->expectExceptionMessageMatches('/You can disable deprecation warnings by setting `error_reporting\(\)` to `E_ALL & ~E_USER_DEPRECATED`\.$/');
-        deprecationWarning('This method is deprecated');
+        $this->deprecated(function () {
+            deprecationWarning('This method is deprecated');
+        });
     }
 }